// src/TeslaClient/Vehicle/protobuf/outputs/universal_message.ts
export enum Domain {
  DOMAIN_BROADCAST = 0,
  DOMAIN_VEHICLE_SECURITY = 2,
  DOMAIN_INFOTAINMENT = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum OperationStatusE {
  OPERATIONSTATUS_OK = 0,
  OPERATIONSTATUS_WAIT = 1,
  OPERATIONSTATUS_ERROR = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface Destination {
  domain?: Domain | undefined;
  routingAddress?: Uint8Array | undefined;
}

export interface MessageStatus {
  operationStatus: OperationStatusE;
  signedMessageFault: MessageFaultE;
}

export interface SessionInfoRequest {
  publicKey: Uint8Array;
  challenge?: Uint8Array | undefined;
}

export interface RoutableMessage {
  toDestination: Destination | undefined;
  fromDestination: Destination | undefined;
  protobufMessageAsBytes?: Uint8Array | undefined;
  sessionInfoRequest?: SessionInfoRequest | undefined;
  sessionInfo?: Uint8Array | undefined;
  signatureData?: SignatureData | undefined;
  signedMessageStatus?: MessageStatus | undefined;
  requestUuid?: Uint8Array | undefined;
  uuid?: Uint8Array | undefined;
  flags?: number | undefined;
}

function createBaseDestination(): Destination {
  return { domain: undefined, routingAddress: undefined };
}

export const Destination: MessageFns<Destination> = {
  encode(message: Destination, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.domain !== undefined) {
      writer.uint32(8).int32(message.domain);
    }
    if (message.routingAddress !== undefined) {
      writer.uint32(18).bytes(message.routingAddress);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Destination {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDestination();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.domain = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.routingAddress = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Destination {
    return {
      domain: isSet(object.domain) ? domainFromJSON(object.domain) : undefined,
      routingAddress: isSet(object.routingAddress) ? bytesFromBase64(object.routingAddress) : undefined,
    };
  },

  toJSON(message: Destination): unknown {
    const obj: any = {};
    if (message.domain !== undefined) {
      obj.domain = domainToJSON(message.domain);
    }
    if (message.routingAddress !== undefined) {
      obj.routingAddress = base64FromBytes(message.routingAddress);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Destination>, I>>(base?: I): Destination {
    return Destination.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Destination>, I>>(object: I): Destination {
    const message = createBaseDestination();
    message.domain = object.domain ?? undefined;
    message.routingAddress = object.routingAddress ?? undefined;
    return message;
  },
};

function createBaseMessageStatus(): MessageStatus {
  return { operationStatus: 0, signedMessageFault: 0 };
}

export const MessageStatus: MessageFns<MessageStatus> = {
  encode(message: MessageStatus, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.operationStatus !== 0) {
      writer.uint32(8).int32(message.operationStatus);
    }
    if (message.signedMessageFault !== 0) {
      writer.uint32(16).int32(message.signedMessageFault);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MessageStatus {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMessageStatus();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.operationStatus = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.signedMessageFault = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MessageStatus {
    return {
      operationStatus: isSet(object.operationStatus) ? operationStatusEFromJSON(object.operationStatus) : 0,
      signedMessageFault: isSet(object.signedMessageFault) ? messageFaultEFromJSON(object.signedMessageFault) : 0,
    };
  },

  toJSON(message: MessageStatus): unknown {
    const obj: any = {};
    if (message.operationStatus !== 0) {
      obj.operationStatus = operationStatusEToJSON(message.operationStatus);
    }
    if (message.signedMessageFault !== 0) {
      obj.signedMessageFault = messageFaultEToJSON(message.signedMessageFault);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MessageStatus>, I>>(base?: I): MessageStatus {
    return MessageStatus.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MessageStatus>, I>>(object: I): MessageStatus {
    const message = createBaseMessageStatus();
    message.operationStatus = object.operationStatus ?? 0;
    message.signedMessageFault = object.signedMessageFault ?? 0;
    return message;
  },
};

function createBaseSessionInfoRequest(): SessionInfoRequest {
  return { publicKey: new Uint8Array(0), challenge: undefined };
}

export const SessionInfoRequest: MessageFns<SessionInfoRequest> = {
  encode(message: SessionInfoRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.publicKey.length !== 0) {
      writer.uint32(10).bytes(message.publicKey);
    }
    if (message.challenge !== undefined) {
      writer.uint32(18).bytes(message.challenge);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SessionInfoRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSessionInfoRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.publicKey = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.challenge = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SessionInfoRequest {
    return {
      publicKey: isSet(object.publicKey) ? bytesFromBase64(object.publicKey) : new Uint8Array(0),
      challenge: isSet(object.challenge) ? bytesFromBase64(object.challenge) : undefined,
    };
  },

  toJSON(message: SessionInfoRequest): unknown {
    const obj: any = {};
    if (message.publicKey.length !== 0) {
      obj.publicKey = base64FromBytes(message.publicKey);
    }
    if (message.challenge !== undefined) {
      obj.challenge = base64FromBytes(message.challenge);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SessionInfoRequest>, I>>(base?: I): SessionInfoRequest {
    return SessionInfoRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SessionInfoRequest>, I>>(object: I): SessionInfoRequest {
    const message = createBaseSessionInfoRequest();
    message.publicKey = object.publicKey ?? new Uint8Array(0);
    message.challenge = object.challenge ?? undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}