// src/TeslaClient/Vehicle/protobuf/outputs/car_server.ts
export interface Action {
  vehicleAction?: VehicleAction | undefined;
}

export interface VehicleAction {
  chargingSetLimitAction?: ChargingSetLimitAction | undefined;
  chargingStartStopAction?: ChargingStartStopAction | undefined;
  drivingClearSpeedLimitPinAction?: DrivingClearSpeedLimitPinAction | undefined;
  drivingSetSpeedLimitAction?: DrivingSetSpeedLimitAction | undefined;
  drivingSpeedLimitAction?: DrivingSpeedLimitAction | undefined;
  hvacAutoAction?: HvacAutoAction | undefined;
  hvacSetPreconditioningMaxAction?: HvacSetPreconditioningMaxAction | undefined;
  hvacSteeringWheelHeaterAction?: HvacSteeringWheelHeaterAction | undefined;
  hvacTemperatureAdjustmentAction?: HvacTemperatureAdjustmentAction | undefined;
  mediaPlayAction?: MediaPlayAction | undefined;
  mediaUpdateVolume?: MediaUpdateVolume | undefined;
  mediaNextFavorite?: MediaNextFavorite | undefined;
  mediaPreviousFavorite?: MediaPreviousFavorite | undefined;
  mediaNextTrack?: MediaNextTrack | undefined;
  mediaPreviousTrack?: MediaPreviousTrack | undefined;
  getNearbyChargingSites?: GetNearbyChargingSites | undefined;
  vehicleControlCancelSoftwareUpdateAction?: VehicleControlCancelSoftwareUpdateAction | undefined;
  vehicleControlFlashLightsAction?: VehicleControlFlashLightsAction | undefined;
  vehicleControlHonkHornAction?: VehicleControlHonkHornAction | undefined;
  vehicleControlResetValetPinAction?: VehicleControlResetValetPinAction | undefined;
  vehicleControlScheduleSoftwareUpdateAction?: VehicleControlScheduleSoftwareUpdateAction | undefined;
  vehicleControlSetSentryModeAction?: VehicleControlSetSentryModeAction | undefined;
  vehicleControlSetValetModeAction?: VehicleControlSetValetModeAction | undefined;
  vehicleControlSunroofOpenCloseAction?: VehicleControlSunroofOpenCloseAction | undefined;
  vehicleControlTriggerHomelinkAction?: VehicleControlTriggerHomelinkAction | undefined;
  vehicleControlWindowAction?: VehicleControlWindowAction | undefined;
  hvacBioweaponModeAction?: HvacBioweaponModeAction | undefined;
  hvacSeatHeaterActions?: HvacSeatHeaterActions | undefined;
  scheduledChargingAction?: ScheduledChargingAction | undefined;
  scheduledDepartureAction?: ScheduledDepartureAction | undefined;
  setChargingAmpsAction?: SetChargingAmpsAction | undefined;
  hvacClimateKeeperAction?: HvacClimateKeeperAction | undefined;
  ping?: Ping | undefined;
  autoSeatClimateAction?: AutoSeatClimateAction | undefined;
  hvacSeatCoolerActions?: HvacSeatCoolerActions | undefined;
  setCabinOverheatProtectionAction?: SetCabinOverheatProtectionAction | undefined;
  setVehicleNameAction?: SetVehicleNameAction | undefined;
  chargePortDoorClose?: ChargePortDoorClose | undefined;
  chargePortDoorOpen?: ChargePortDoorOpen | undefined;
  guestModeAction?: VehicleState_GuestMode | undefined;
  setCopTempAction?: SetCopTempAction | undefined;
  eraseUserDataAction?: EraseUserDataAction | undefined;
  vehicleControlSetPinToDriveAction?: VehicleControlSetPinToDriveAction | undefined;
  vehicleControlResetPinToDriveAction?: VehicleControlResetPinToDriveAction | undefined;
  addChargeScheduleAction?: ChargeSchedule | undefined;
  removeChargeScheduleAction?: RemoveChargeScheduleAction | undefined;
  addPreconditionScheduleAction?: PreconditionSchedule | undefined;
  removePreconditionScheduleAction?: RemovePreconditionScheduleAction | undefined;
  batchRemovePreconditionSchedulesAction?: BatchRemovePreconditionSchedulesAction | undefined;
  batchRemoveChargeSchedulesAction?: BatchRemoveChargeSchedulesAction | undefined;
}
// ...
export interface ChargingSetLimitAction {
  percent: number;
}

export interface ChargingStartStopAction {
  unknown?: Void | undefined;
  start?: Void | undefined;
  startStandard?: Void | undefined;
  startMaxRange?: Void | undefined;
  stop?: Void | undefined;
}

export interface DrivingClearSpeedLimitPinAction {
  pin: string;
}

export interface DrivingSetSpeedLimitAction {
  limitMph: number;
}

export interface DrivingSpeedLimitAction {
  activate: boolean;
  pin: string;
}

export interface HvacAutoAction {
  powerOn?: boolean | undefined;
  manualOverride?: boolean | undefined;
}
// ...
export interface HvacSteeringWheelHeaterAction {
  powerOn: boolean;
}
// ...
export const Action: MessageFns<Action> = {
  encode(message: Action, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.vehicleAction !== undefined) {
      VehicleAction.encode(message.vehicleAction, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Action {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAction();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.vehicleAction = VehicleAction.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Action {
    return { vehicleAction: isSet(object.vehicleAction) ? VehicleAction.fromJSON(object.vehicleAction) : undefined };
  },

  toJSON(message: Action): unknown {
    const obj: any = {};
    if (message.vehicleAction !== undefined) {
      obj.vehicleAction = VehicleAction.toJSON(message.vehicleAction);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Action>, I>>(base?: I): Action {
    return Action.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Action>, I>>(object: I): Action {
    const message = createBaseAction();
    message.vehicleAction = (object.vehicleAction !== undefined && object.vehicleAction !== null)
      ? VehicleAction.fromPartial(object.vehicleAction)
      : undefined;
    return message;
  },
};
// ...
export const ChargingSetLimitAction: MessageFns<ChargingSetLimitAction> = {
  encode(message: ChargingSetLimitAction, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.percent !== 0) {
      writer.uint32(8).int32(message.percent);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChargingSetLimitAction {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChargingSetLimitAction();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.percent = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ChargingSetLimitAction {
    return { percent: isSet(object.percent) ? globalThis.Number(object.percent) : 0 };
  },

  toJSON(message: ChargingSetLimitAction): unknown {
    const obj: any = {};
    if (message.percent !== 0) {
      obj.percent = Math.round(message.percent);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ChargingSetLimitAction>, I>>(base?: I): ChargingSetLimitAction {
    return ChargingSetLimitAction.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChargingSetLimitAction>, I>>(object: I): ChargingSetLimitAction {
    const message = createBaseChargingSetLimitAction();
    message.percent = object.percent ?? 0;
    return message;
  },
};
// ...
export const DrivingClearSpeedLimitPinAction: MessageFns<DrivingClearSpeedLimitPinAction> = {
  encode(message: DrivingClearSpeedLimitPinAction, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.pin !== "") {
      writer.uint32(10).string(message.pin);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DrivingClearSpeedLimitPinAction {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDrivingClearSpeedLimitPinAction();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.pin = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DrivingClearSpeedLimitPinAction {
    return { pin: isSet(object.pin) ? globalThis.String(object.pin) : "" };
  },

  toJSON(message: DrivingClearSpeedLimitPinAction): unknown {
    const obj: any = {};
    if (message.pin !== "") {
      obj.pin = message.pin;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DrivingClearSpeedLimitPinAction>, I>>(base?: I): DrivingClearSpeedLimitPinAction {
    return DrivingClearSpeedLimitPinAction.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DrivingClearSpeedLimitPinAction>, I>>(
    object: I,
  ): DrivingClearSpeedLimitPinAction {
    const message = createBaseDrivingClearSpeedLimitPinAction();
    message.pin = object.pin ?? "";
    return message;
  },
};

function createBaseDrivingSetSpeedLimitAction(): DrivingSetSpeedLimitAction {
  return { limitMph: 0 };
}

export const DrivingSetSpeedLimitAction: MessageFns<DrivingSetSpeedLimitAction> = {
  encode(message: DrivingSetSpeedLimitAction, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.limitMph !== 0) {
      writer.uint32(9).double(message.limitMph);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DrivingSetSpeedLimitAction {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDrivingSetSpeedLimitAction();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 9) {
            break;
          }

          message.limitMph = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DrivingSetSpeedLimitAction {
    return { limitMph: isSet(object.limitMph) ? globalThis.Number(object.limitMph) : 0 };
  },

  toJSON(message: DrivingSetSpeedLimitAction): unknown {
    const obj: any = {};
    if (message.limitMph !== 0) {
      obj.limitMph = message.limitMph;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DrivingSetSpeedLimitAction>, I>>(base?: I): DrivingSetSpeedLimitAction {
    return DrivingSetSpeedLimitAction.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DrivingSetSpeedLimitAction>, I>>(object: I): DrivingSetSpeedLimitAction {
    const message = createBaseDrivingSetSpeedLimitAction();
    message.limitMph = object.limitMph ?? 0;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}